local_testing_scripts/requirements.txt
boto3>=1.34
Pillow>=10.0
